Allow tested controllers to be embedded in standard containers
When declaring the controller under test, standard iOS container controllers
can now be stipulated through keyword options:

1. :navigation (UINavigationController)
2. :tab        (UITabBarController)
3. :present    (-presentViewController:animated:completion:)

Example:

    tests HamsterdamController, navigation: true

The use of container controllers is concatenative. For example, if both
:navigation and :tab are specified, the controller under test will be first
embedded in a navigation controller, and the navigation controller will be
embedded in a tab controller.

In the tests, #controller continues refers to the instance of the controller
under test even though it is no longer the window's root view controller. To
access a container controller, use one of UIViewControllers related controller
methods, such as #navigationController.